Refector Liquidation interface and optimices Params Manager

+++ database/migrations/2019_05_14_193946_create_bea_marks_table.php
@@ -26,8 +26,8 @@ class CreateBeaMarksTable extends Migration
             $table->integer('locks');
             $table->integer('auxiliaries');
             $table->integer('boarded');
-            $table->integer('im_bea_max');
-            $table->integer('im_bea_min');
+            $table->double('im_bea_max');
+            $table->double('im_bea_min');
             $table->integer('total_bea');
             $table->integer('passengers_bea');
